--- a/test/perf/package.json
+++ b/test/perf/package.json
@@ -1,0 +1,23 @@
+{
+    "name": "@capire/sflight-performance",
+    "version": "1.0.0",
+    "private": true,
+    "description": "CAP flight demo scenario performance tests",
+    "license": "SAP SAMPLE CODE LICENSE",
+    "repository": "https://github.com/SAP-samples/cap-sflight",
+    "engines": {
+        "node": ">=16"
+    },
+    "scripts": {
+        "start": "node index.js"
+    },
+    "dependencies": {
+        "@sap/cds": ">=7.0.0",
+        "axios": "^1",
+        "jest": "^29.0.2"
+    },
+    "jest": {
+        "testEnvironment": "node",
+        "testTimeout": 3600000
+    }
+}