--- a/test/perf/srv/analytics-service.cds
+++ b/test/perf/srv/analytics-service.cds
@@ -1,0 +1,1 @@
+using from '../../../srv/analytics-service';