--- a/test/perf/srv/travel-service.cds
+++ b/test/perf/srv/travel-service.cds
@@ -1,0 +1,1 @@
+using from '../../../srv/travel-service';