--- a/test/perf/srv/travel-service.js
+++ b/test/perf/srv/travel-service.js
@@ -1,0 +1,1 @@
+module.exports = require('../../../srv/travel-service')