--- a/test/perf/srv/workarounds.js
+++ b/test/perf/srv/workarounds.js
@@ -1,0 +1,1 @@
+module.exports = require('../../../srv/workarounds')