@BobdenOs 2024-03-21
Local performance run with all available protocol adapters and most database service.

--------------------------------
### landingPageList
<details>
<summary>baseline (1 users)</summary>

|   |sqlite (new)|sqlite disk (new)|sqlite (old)|sqlite disk (old)|hana hdb (new)|hana hana-client (new)|postgres|
|---|---|---|---|---|---|---|---|
| okra | 5 ms (avg: 7 ms) |  6 ms (avg: 7 ms) |  14 ms (avg: 16 ms) |  14 ms (avg: 16 ms) |  87 ms (avg: 91 ms) |  85 ms (avg: 93 ms) |  14 ms (avg: 19 ms) | 
| odata | 5 ms (avg: 7 ms) |  5 ms (avg: 7 ms) |  14 ms (avg: 15 ms) |  14 ms (avg: 15 ms) |  87 ms (avg: 92 ms) |  85 ms (avg: 90 ms) |  15 ms (avg: 19 ms) | 
| rest | 5 ms (avg: 6 ms) |  5 ms (avg: 6 ms) |  13 ms (avg: 15 ms) |  13 ms (avg: 15 ms) |  86 ms (avg: 91 ms) |  83 ms (avg: 90 ms) |  13 ms (avg: 18 ms) | 
| graphql | 6 ms (avg: 7 ms) |  5 ms (avg: 7 ms) |  10 ms (avg: 12 ms) |  10 ms (avg: 12 ms) |  84 ms (avg: 89 ms) |  84 ms (avg: 88 ms) |  13 ms (avg: 18 ms) | 
| hcql | 7 ms (avg: 8 ms) |  7 ms (avg: 8 ms) |  15 ms (avg: 16 ms) |  14 ms (avg: 16 ms) |  62 ms (avg: 68 ms) |  60 ms (avg: 65 ms) |  17 ms (avg: 21 ms) | 
</details>
<details>
<summary>m (128 users)</summary>

|   |sqlite (new)|sqlite disk (new)|sqlite (old)|sqlite disk (old)|hana hdb (new)|hana hana-client (new)|postgres|
|---|---|---|---|---|---|---|---|
| okra | 2 ms (avg: 4 ms) |  2 ms (avg: 4 ms) |  10 ms (avg: 13 ms) |  9 ms (avg: 13 ms) |  12 ms (avg: 76 ms) |  11 ms (avg: 76 ms) |  8 ms (avg: 15 ms) | 
| odata | 2 ms (avg: 4 ms) |  2 ms (avg: 4 ms) |  10 ms (avg: 13 ms) |  10 ms (avg: 13 ms) |  11 ms (avg: 75 ms) |  12 ms (avg: 77 ms) |  9 ms (avg: 15 ms) | 
| rest | 3 ms (avg: 5 ms) |  3 ms (avg: 4 ms) |  10 ms (avg: 14 ms) |  10 ms (avg: 14 ms) |  12 ms (avg: 95 ms) |  12 ms (avg: 99 ms) |  9 ms (avg: 19 ms) | 
| graphql | 3 ms (avg: 5 ms) |  3 ms (avg: 5 ms) |  6 ms (avg: 10 ms) |  6 ms (avg: 10 ms) |  12 ms (avg: 959 ms) |  13 ms (avg: 80 ms) |  7 ms (avg: 14 ms) | 
| hcql | 4 ms (avg: 6 ms) |  4 ms (avg: 6 ms) |  11 ms (avg: 15 ms) |  11 ms (avg: 14 ms) |  20 ms (avg: 593 ms) |  20 ms (avg: 557 ms) |  10 ms (avg: 29 ms) | 
</details>